# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: LibreRun
  version: 1.1.0-beta.1
paths:
...
  /api/v1/runs:
    get:
      tags:
      - runs
      summary: List Runs
      operationId: list_runs_api_v1_runs_get
      security:
      - HTTPBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: severity
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Severity
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: per_page
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Per Page
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - runs
      summary: Create Run
      description: 'Create a run from a schema-validated payload (blueprint B8).


        The body shape belongs to the selected agent: it is validated against

        ``agent.input_schema()`` server-side (the wizard enforces the same

        schema client-side), ``x-pii`` string fields are redacted before

        anything is persisted, and well-known keys are lifted into the legacy

        list/search columns.'
      operationId: create_run_api_v1_runs_post
      security:
      - HTTPBearer: []
      parameters:
      - name: agent_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Agent Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              title: Payload
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunCreatedResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    RunCreatedResponse:
      properties:
        run_id:
          type: string
          format: uuid
          title: Run Id
        run_number:
          type: string
          title: Run Number
        status:
          type: string
          title: Status
        case_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Case Id
          description: 'Deprecated duplicate of `run_id` (LibreRun speaks `run`, blueprint
            S1 / decision L18): always identical to `run_id`, emitted for one release
            and removed at v1.1. Read `run_id` instead.'
          deprecated: true
        case_number:
          anyOf:
          - type: string
          - type: 'null'
          title: Case Number
          description: 'Deprecated duplicate of `run_number` (LibreRun speaks `run`,
            blueprint S1 / decision L18): always identical to `run_number`, emitted
            for one release and removed at v1.1. Read `run_number` instead.'
          deprecated: true
      type: object
      required:
      - run_id
      - run_number
      - status
      title: RunCreatedResponse
...
    RunListResponse:
      properties:
        runs:
          items:
            $ref: '#/components/schemas/RunSummary'
          type: array
          title: Runs
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        per_page:
          type: integer
          title: Per Page
      type: object
      required:
      - runs
      - total
      - page
      - per_page
      title: RunListResponse
    RunSummary:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        run_number:
          type: string
          title: Run Number
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        vendor_a_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Vendor A Name
        vendor_b_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Vendor B Name
        problem_summary:
          type: string
          title: Problem Summary
          description: 'Deprecated (LibreRun blueprint S2): `title` replaces it; kept
            identical for one release and removed at v1.1. Read `title` instead.'
          default: ''
          deprecated: true
        severity:
          anyOf:
          - type: string
          - type: 'null'
          title: Severity
        status:
          type: string
          title: Status
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        agent_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Agent Id
        case_number:
          anyOf:
          - type: string
          - type: 'null'
          title: Case Number
          description: 'Deprecated duplicate of `run_number` (LibreRun speaks `run`,
            blueprint S1 / decision L18): always identical to `run_number`, emitted
            for one release and removed at v1.1. Read `run_number` instead.'
          deprecated: true
      type: object
      required:
      - id
      - run_number
      - severity
      - status
      - created_at
      - updated_at
      title: RunSummary
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer